previous changes reverted, check for contents of opioid file

+++ main.sh
@@ -12,45 +12,30 @@ readonly VENV_PATH="env"
 
 main() {
 
-    if pip --version >/dev/null 2>&1; then
-        echo "pip already installed"
-    else
-        echo "installing pip..."
-        #python -m pip install
-        #sudo easy_install pip
-        curl https://bootstrap.pypa.io/get-pip.py -o get-pip.py
-        python get-pip.py
-        echo "pip installed"
-    fi
-
-    if virtualenv --version >/dev/null 2>&1; then
-        echo "virtualenv already installed"
-    else
-        echo "Installing virtuaalenv"
-        pip install virtualenv
-        echo "virtualenv installed"
-    fi
+    echo "Installing virtualenv"
+    pip install virtualenv
+    echo "Done installing virtualenv"
 
     if [ -d $VENV_PATH ]; then
         echo "Skipping venv creation as it already exists."
     else
-        echo "Creating venv..."
+        echo "Creating virtualenv..."
         python -m virtualenv $VENV_PATH
-        echo "Done Creating venv"
+        echo "Done Creating virtualenv"
     fi
 
     #activate venv
     source $VENV_PATH/bin/activate
-    pip install -r requirements.txt #put this in an if block
+    pip install -r requirements.txt
     echo "In venv"
 
-    #if [ -f $DATASET1_PATH ]; then
-    #    echo "Skipping download since opioid dataset exists"
-    #else
-    echo "Downloading opioid dataset..."
-    curl -o $DATASET1_PATH https://data.ct.gov/api/views/rybz-nyjw/rows.csv?accessType=DOWNLOAD
-    echo "Done downloading opioid dataset"
-    #fi
+    if [ -s $DATASET1_PATH ]; then
+        echo "Skipping download since opioid dataset exists"
+    else
+        echo "Downloading opioid dataset..."
+        curl -o $DATASET1_PATH https://data.ct.gov/api/views/rybz-nyjw/rows.csv?accessType=DOWNLOAD
+        echo "Done downloading opioid dataset"
+    fi
 
     python Main.py
